+++ CHANGELOG.md
@@ -8,18 +8,63 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Fixed
+### Added
 
-- Don't panic when AWSControlPlane CR AZs are nil.
+- Add etcd client certificates for Prometheus.
 
 ### Changed
 
-- Updated workflows for automatic release.
-- Updated backward incompatible Kubernetes dependencies to v1.18.5.
+- Update `k8scloudconfig` to `v8.0.3`.
+
+
+## [9.1.1] - 2020-09-23
 
 ### Fixed
 
+- Update flatcar AMI for China
+
+## [9.1.0] - 2020-09-22
+
+- Update AWS CNI manifests
+- Disable Calico CNI binaries installation
+
+## [9.0.1] - 2020-09-17
+
+- Update flatcar releases
+
+## [9.0.0] - 2020-09-15
+
+### Added
+
+- Emit Kubernetes events for tcnpf Cloudformation stack failures
+- Emit Kubernetes events for tccpi and tccpf Cloudformation stack failures
+- Add monitoring label
+- Handle the case when there are both public and private hosted zones for CP
+  base domain.
+- Add Route Table lookup using tags, so `RouteTables` flag can be phased out in the future.
+
+
+### Changed
+
+- Update backward incompatible Kubernetes dependencies to v1.18.5.
+- Remove migration code to ensure the Control Plane CRs for existing Node Pool clusters.
+
+### Deprecated
+
+- `RouteTables` flag will be deprecated.
+
+### Fixed
+
+- Don't panic when AWSControlPlane CR AZs are nil.
+- Add suffix to Route Tables to get rid of naming collision.
 - Fix image-pull-progress-deadline argument for tcnp nodes.
+
+### Removed
+
+- Remove etcd snapshot migration code.
+- Remove unused `--service.aws.accesskey.id`, `--service.aws.accesskey.secret`
+  and `--service.aws.accesskey.session` flags.
+- Remove the prometheus collector and move it to the separate `aws-collector` project.
 
 ## [8.8.0] - 2020-08-14
 
@@ -230,7 +275,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v8.8.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v9.1.1...HEAD
+[9.1.1]: https://github.com/giantswarm/aws-operator/compare/v9.1.0...v9.1.1
+[9.1.0]: https://github.com/giantswarm/aws-operator/compare/v9.0.1...v9.1.0
+[9.0.1]: https://github.com/giantswarm/aws-operator/compare/v9.0.0...v9.0.1
+[9.0.0]: https://github.com/giantswarm/aws-operator/compare/v8.8.0...v9.0.0
 [8.8.0]: https://github.com/giantswarm/aws-operator/compare/v8.7.6...v8.8.0
 [8.7.6]: https://github.com/giantswarm/aws-operator/compare/v8.7.5...v8.7.6
 [8.7.5]: https://github.com/giantswarm/aws-operator/compare/v8.7.4...v8.7.5